lps22hh: fix Function fix         _from_lsb_to_hpa

+++ lps22hh_STdC/driver/lps22hh_reg.c
@@ -100,7 +100,7 @@ int32_t lps22hh_write_reg(lps22hh_ctx_t* ctx, uint8_t reg, uint8_t* data,
   * @{
   *
   */
-float_t lps22hh_from_lsb_to_hpa(int16_t lsb)
+float_t lps22hh_from_lsb_to_hpa(int32_t lsb)
 {
   return ( (float_t) lsb / 4096.0f );
 }
